+++ app/src/main/java/com/mezyapps/vgreenbasket/db/CardAppDAO.java
@@ -40,5 +40,5 @@ public interface CardAppDAO {
     public long getProductQtyUpdate(long qty,long product_id);
 
     @Query("UPDATE ProductTable SET qty = :qty,price_TOTAl = :price_total,MRP_TOTAL = :mrp_total WHERE product_id = :product_id")
-    public long getProductDataUpdate(long qty,long product_id,long price_total,long mrp_total);
+    public long getProductDataUpdate(long qty,long product_id,double price_total,double mrp_total);
 }
